requirements.txt
PyQt6==6.7.1 
Pillow==10.4.0 
lxml==5.3.0 
markdown==3.7 
PyPDF2==3.0.1 
pdfplumber==0.11.4 
PyMuPDF==1.24.12 
httpx==0.27.2